Make items objects instead of arrays for tuples
Workaround for https://github.com/biocad/openapi3/issues/31

If the tuples has homogeneous types, the generated schema is strict.
On the other hand if there are heterogeneous types, the schema is not very strict because the order in which the types must come is not specified.

Also, I had to use anyOf instead of oneOf because for example the int in (Int, Float) matches both Integer and Number.

Finally, special care had to be taken to handle nullables.